Anim: Add Action+Slot selectors to Property editor tabs
Add Action + Action Slot selectors to various Property Editor tabs.
This follows the pattern established in
f917b60036976a2fe2e7f990e171f20cb502a261.

Support for `CacheFile` data-blocks is not included here. That needs
some discussion, as currently its interface is part of the Modifier
stack (and thus implemented in C++ and not Python, and using a
different layout).

Since this PR is about adding to the Property Editor, some data-blocks
that have no representation there are thus excluded (like `Mask`).
Even when they could actually use an Action+Slot selector somewhere.

Pull Request: https://projects.blender.org/blender/blender/pulls/127074

+++ scripts/startup/bl_ui/space_properties.py
@@ -112,8 +112,15 @@ class PropertiesAnimationMixin:
         col.use_property_split = True
         self.draw_action_and_slot_selector(context, col, self._animated_id(context))
 
-    @staticmethod
-    def draw_action_and_slot_selector(context, layout, animated_id):
+    @classmethod
+    def draw_action_and_slot_selector(cls, context, layout, animated_id):
+        if not animated_id:
+            class_list = [c.__name__ for c in cls.mro()]
+            print("PropertiesAnimationMixin: no animatable data-block, this is a bug "
+                  "in one of these classes: {}".format(class_list))
+            layout.label(text='No animatable data-block, please report as bug', icon='ERROR')
+            return
+
         layout.template_action(animated_id, new="action.new", unlink="action.unlink")
 
         if not context.preferences.experimental.use_animation_baklava:
